Replace paralellism with suspend functions
All jobs are now shared by the Kotlin-coroutines common pool of
threads. Crucially this means that blocking in a worker function risks
blocking all message processing. On work loads with great variance
time taken between tasks it is recomended offload the larger jobs to a
separate pool so as to ensure that smaller jobs can still be processed