Do not store WorkerInterceptor(s) in a list
At setup phase of WorkerInterceptors, they are stored in a list, even tough they
are by them self a linked list. Now this is utilized from start, which means that we
can use copy-on-write when we modify this list, and get rid of the mutability of this
list that caused a race condition. This is an alternate solution on the NPE that @bonahona
has adressed in pull request #34.