Use PriorityBlockingQueues in KafkaConsumerActor
Use PriorityBlockingQueues in KafkaConsumerActor, instead of regular
ArrayBlockingQueues, so messages can be sorted by priority. This should
allow Franz to prioritize previously failed messages before new
messages. The heuristic for ordering messages is roughly as follows

- Prioritize messages that are failed before new messages
- Prioritize messages with a lower offset over messages with a higher
offset (given that they belong to the same topic and partition
- Prioritize older messages over newer messages (by Kafka timestamp)

This functionality is yet to be tested, and is not verified working at
this point.